Extract buildEndPointUrl() into a UrlUtil class
