+++ src/main/java/org/gridsuite/study/server/StudyConstants.java
@@ -30,7 +30,7 @@ public final class StudyConstants {
     public static final String REPORT_API_VERSION = "v1";
     public static final String SHORT_CIRCUIT_API_VERSION = "v1";
     public static final String TIME_SERIES_API_VERSION = "v1";
-    public static final String DYNAMIC_MAPPING_API_VERSION = "v1";
+    public static final String DYNAMIC_MAPPING_API_VERSION = "";
 
     public static final String NETWORK_UUID = "networkUuid";
     public static final String CASE_UUID = "caseUuid";
